Address code review feedback
