+++ src/vm/gcenv.ee.cpp
@@ -13,7 +13,11 @@
 
 #include "common.h"
 
+#ifdef FEATURE_STANDALONE_GC
 #include "gcenv.ee.h"
+#else
+#include "../gc/env/gcenv.ee.h"
+#endif // FEATURE_STANDALONE_GC
 
 #ifdef FEATURE_COMINTEROP
 #include <windows.ui.xaml.h>
